requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
celery==5.3.1
redis==4.6.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Jinja2==3.1.2
email-validator==2.0.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
flask-cors==4.0.0